CCEffectBlur - Experiment with the linearly filtered gaussian blur logic
This gets a simple filter (though not actually a gaussian distribution) working the same
way with and without the linear sampling optimization. Now to figure out why it works
here but does not work in the real code.  The weights and offsets are correct in the
real code so I think the problem is in the conversion of the offsets to texture coordinates.

When I say "correct" in the previous paragraph that is to say that if you choose a pair of
samples in the texture and compute their weighted combination by hand with and without
the linear filtering optimization you get the same result.