CCEffects - Move bounded texture sampling code into shared location
Effects that sample outside the current pixel all share some code that determines whether or
not they're sampling outside the bounds of the associated texture (for example, this can happen
when blurring a sprite from a sprite sheet and padding is applied). In the past this code was inlined
everywhere it was needed but now it's moved into a single shared location within the shader
builders.