CCEffectDropShadow - Refactor to use the shader from CCEffectBlur
Drop shadow is just three passes: two to implement the separable blur and one
to composite the blurred shadow with the original sprite. Instead of reimplementing
all the code from CCEffectBlur for those blur passes, take advantage of our ability
to have multiple shaders per-effect and directly use the shader from CCEffectBlur.

This also includes fixes to the GL stitcher to handle stitching of effects with multiple
shaders. This was broken.